Trailing path separator in g:snippets_dir causes problems with globpath() on Windows
A trailing backslash as path separator causes problems when multiple directories are enumerated: The comma is treated as escaped, and the entire enumeration is taken as one (nonexisting) dirspec.
To avoid this, drop the trailing separator when generating g:snippets_dir; it's actually not needed elsewhere in the code, anyway.