Removed: All default snippets except for snippets itself
Some of those I have never used, and the others I have to maintain in this (mostly dead) fork, distracting from the actual plugin maintenance. Also, these customizations are my private ones and not useful to others who might want to use my fork for the plugin functionality.
So, move away the snippets I privately use to ~/.vim/snippets, and remove the rest.